Number destination upon arrival
Added optional arguments to OSRMInstructionFormatter.string(for:modifyValueByKey:) to track which leg the instruction is on. When there are multiple waypoints in the route, indicate the ordinal number of the waypoint in the arrival instruction.

Fixed the raw value of TokenType.wayPoint.

Updated tests. Ensure that formatted ordinals match the fixture language (English) rather than the test machine’s language.